Add shp for Gujarat AC.
NB there are some big chunks of Gujarat missing.
It appears these are not covered by any constituency.
This is corroborated by:
https://upload.wikimedia.org/wikipedia/commons/3/3e/Wahlkreise_zur_Vidhan_Sabha_von_Gujarat.svg
and
https://timesofindia.indiatimes.com/elections/assembly-elections/gujarat/results
An official source has not been found, since the delimitation only
records the areas that make up a constituency, not those that
are not covered. At any rate a lot of the missing chunks look
like desert.